in_tail: Add group based log collection from `in_tail_with_throttle`
